Allow included application to create versions
If FedoraMigrate is included in another application that is already
creating versions for attached files, then FedoraMigrate should not
create versions when migrating. Otherwise, there are two versions for
every one original version. Here we pass an option that lets the
included application create the version. By default, however,
FedoraMigrate will create the versions such as if it is used outside
the context of a Hydra application, or one in which no versions are
automatically created.